Add QMacPasteboardMime.
Equivalent to QMacPasteboardMime from Qt 4. Requires
exported functions from QPlatformNativeInterface.

The API from Qt 4 is in large parts kept, with the
exception of the static functions used internally
by Qt (?)

The implementation is based on having two equal classes
and keeping them in sync: One internal to the Cocoa 
platform plugin and one public in QtMacExtras.

Change-Id: I334fc982355fbb6315d7a44eeca8551071ba60be