Add QtMacPastebardMime.
Replaces QMacPasteboardMime from Qt 5. Requires
Qt 5.0.1 which exports new (private) functions from
QPlatformNativeInterface.

The API from Qt 4 is in large parts kept, with the
exception of the static functions used internally
by Qt (?) 

The implementation is based on having two classes and
keeping them in sync. QMacPasteboardMime is internal
to the Cocoa platform plugin. QtMacPastebardMime is
public in QtMacExtras. These can be casted between
such that subclasses of QtMacPastebardMime can be
registered and used by Qt.

Also add qRegisteredDraggedTypes which was exported
in Qt 4.

Change-Id: I334fc982355fbb6315d7a44eeca8551071ba60be