Reintroduce setUnifiedTitleAndToolBarOnMac as a global function.
The function creates a QtMacUnifiedToolBar with the same actions as
a QToolBar, adds the native toolbar to the window and then hides
the original QToolBar. A dummy implementation that does nothing is
provided on non-Mac platforms so no ifdefs are necessary for client
code just like the original function.

Essentially the only changes necessary from Qt 4 to get a real native
toolbar are to change 'true' to a pointer to your QToolBar (and add
a global scope qualifier :: if necessary).

This commit also introduces the ability to remove the
QtMacUnifiedToolBar from a window if it has been added to one.

Change-Id: I8b15a21c5a3850fcb0e1b058a85cecca98048467