Do not use the same identifier for all NSToolBars.
This *will* cause problems if an app has more than one toolbar, such
as method calls to one toolbar affecting all of them, so we'll have
the constructor use a QUuid to generate a unique identifier for each
toolbar created. Cocoa also uses a UUID as the toolbar identifier when
set to auto-generate in Interface Builder.

Also allow the user to specify a custom identifier if they desire.

Change-Id: I7d5c8d58e45a504480a1ce67065add15cb35ad8f