Do not use the same identifier for all NSToolBars.
This will cause problems if an app has more than one toolbar with a
different set of items. When two toolbars have the same identifier,
Cocoa will attempt to synchronize their state - i.e. order of items,
size of icons, button style.

Because you only want this behavior if two toolbars are *actually*
the same, developers must be able to explicitly set the identifier.
For example, all instances of toolbars belonging to a document
window class should have the same identifier as they all have the
same items and are "logically" the same toolbar. However the identifier
of a document window toolbar and a preferences window toolbar for
example, should not be the same as they are not logically related and
probably have different items which will cause a crash if/when Cocoa
attempts their synchronization.

This commit sets the QtMacUnifiedToolBar constructor to use a QUuid
to generate a unique identifier for each toolbar created unless the
developer sets one explicitly. Cocoa also uses a UUID as the toolbar
identifier when set to auto-generate in Interface Builder, so this
is consistent with native behavior.

Change-Id: I7d5c8d58e45a504480a1ce67065add15cb35ad8f
Reviewed-by: Morten Johan Sørvig <morten.sorvig@digia.com>