Fix a couple minor compile issues.
NULL is a preprocessor definition, not a keyword. It might not be
defined, so just use 0 (no point in including cstdlib JUST for that).

Similarly, Q_OS_MAC might not be defined when this header is run through
moc, depending on the environment, nor is there really any reason to
protect this class definition with an ifdef. Users can do that
themselves in their own code, and we may later provide a dummy
implementation for non-Mac platforms anyways.

Change-Id: I92905e5b95c31df99e43f1aa198980e8a6c6268a