Always create a native toolbar if the given QToolBar is not null.
Previously, if the user tried to create a native toolbar but set
on = false we tried to be conservative and not create the NSToolbar
just yet (fairly non-obvious). The user might not want to have the
toolbar displayed immediately but still obtain the pointer for later
use, so we'll support this use case by always returning a toolbar.

Change-Id: I94dcd3ae57f0462893b2539623423f2a2d85437a