Doc: Change instances of 'OS X' to 'macOS'
As of version 10.12 (Sierra), the name of Apple's desktop operating
system will be macOS. Change all occurrences where the Mac platform
is discussed to use the macro \macos (defined in the documentation
configuration in qtbase), except in link targets (QDoc doesn't
expand macros in those).

Add a 'macOS' prefix to example titles to make it clear they are
specific to that platform.

Change-Id: I6bac8d92ca080adf7138a12e0f792c132e382f5d
Reviewed-by: Jake Petroules <jake.petroules@qt.io>
Reviewed-by: Leena Miettinen <riitta-leena.miettinen@qt.io>