Doc: Fix documentation issues when using Clang-enabled QDoc
QDoc in Qt 5.11 uses Clang to parse the C++ source for documentation
comments. Unlike the old parser, Clang requires a module header to
be present as well as correct include paths to generate the docs
correctly. However, these are available from the build system only
when building on macOS, not when generating the documentation
on another system (e.g. Linux).

To fix this, create a 'dummy' module header specific to doc builds
and add required relative include paths and DEFINEs to .qdocconf.

Change-Id: Iadc1475d16c5570b86de7ea8e17bfeb02d8412eb
Reviewed-by: Martin Smith <martin.smith@qt.io>