Deprecate APIs where other alternatives are preferred
Some of the APIs in this module now have APIs in other parts of Qt,
more closely tied to the classes they convert back and forth from.

Other APIs have native equivalents that are just as simple as the
Qt wrapper, and we want people to use the native APIs since that
scales better, and teaches users that they can mix native and Qt
APIs.

Change-Id: Id587cefe3ab8bee9f7d3e7dc4cb223882d57d325
Reviewed-by: Timur Pocheptsov <timur.pocheptsov@qt.io>