Update commercial license headers
Updated header.COMM to the file in tqtc-qtmacextras. Examples,
tests or documentation files are not updated.

The commercial license header may contain some
additional lines so that its line count equals with
the earlier license header. Reason for this is that
some autotests use hard coded line numbers and a
change in the line count causes failures in tests.

Task-number: QTQAINFRA-4169
Change-Id: I6a50040df036bb22491ee8acca04a72e476224dc
Reviewed-by: Antti Kokko <antti.kokko@qt.io>